Refactor NodeIPAMController to include a starter
